Support Vector Machine and GridSearch
In this project we used Breast_cancer dataset and based on several features we are trying to predict if a cancer is malignant or benign. Here, first we created model by SVM and later we created same model by finding best parameters using GridSearch and compared accuracy of both the cases.